refactor(cli): separate next runtime shims
